Refactor iTunes tags for April 2019 changes
There is a number of breaking changes in this commit that is required for the
iTunes tag updates.

There is also a lot of new logic coming (in future commits) that is required to
parse the tags properly.

This is an initial draft.  More plumbing is required, as well as reworking all
Examples and tests.

One bit of good news: lots of new documentation is now available on iTunes. So
I've added a lot of it here.